backend/requirements.txt
Django>=3.0,<4.0
psycopg2>=2.8

django-cors-headers==3.5.0
pika==1.1.0
packaging==16.8
pyparsing==2.2.0
pytz==2017.2
six==1.10.0
appdirs==1.4.3

djoser==2.0.5
djangorestframework==3.11.2
markdown==3.3
django-filter==2.4.0
djangorestframework-simplejwt==5.0.0
PyJWT
django-enumchoicefield >= 2.0.0
Pillow >= 8.0
stripe
unittest-xml-reporting
xmlrunner
coverage
pytest-django
python-environ

mysqlclient==2.1.1
sqlparse==0.4.3
tzdata==2022.7
asgiref==3.5.2
